adding backpropagation

+++ src/activation.rs
@@ -1,5 +1,5 @@
 #[derive(Debug)]
 pub struct Activation {
-    pub function: fn(f32) -> f32,
-    pub derivative: fn(f32) -> f32,
+    pub function: fn(f64) -> f64,
+    pub derivative: fn(f64) -> f64,
 }
