+++ src/layer.rs
@@ -1,3 +1,3 @@
 pub trait Layer {
-    fn get_holded_values(&self) -> impl Iterator<Item = f32>;
+    fn get_holded_values(&self) -> impl Iterator<Item = f64>;
 }
